cursos/utils/telegram.py
# utils/telegram.py
import requests

from config.settings.base import env


def send_telegram_notification(message):
    bot_token = env("TELEGRAM_BOT_TOKEN")
    chat_id = env("TELEGRAM_CHAT_ID")
    url = f"https://api.telegram.org/bot{bot_token}/sendMessage"
    payload = {
        "chat_id": chat_id,
        "text": message,
    }
    response = requests.post(url, json=payload, timeout=5)
    return response.json()
